fix(schema): fixes shcmea being in multiple pieces , added test case to catch the change
currently its duplicated in api and osmsg to keep api image slim
