+++ api/db.py
@@ -3,52 +3,9 @@ import os
 import asyncpg
 from dotenv import load_dotenv
 
+from .pg_schema import PG_SCHEMA
+
 load_dotenv()
-
-
-PG_SCHEMA = """
-CREATE TABLE IF NOT EXISTS users (
-    uid      BIGINT PRIMARY KEY,
-    username TEXT NOT NULL
-);
-CREATE TABLE IF NOT EXISTS changesets (
-    changeset_id BIGINT PRIMARY KEY,
-    uid          BIGINT NOT NULL REFERENCES users(uid),
-    created_at   TIMESTAMPTZ,
-    hashtags     TEXT[],
-    editor       TEXT,
-    min_lon      DOUBLE PRECISION,
-    min_lat      DOUBLE PRECISION,
-    max_lon      DOUBLE PRECISION,
-    max_lat      DOUBLE PRECISION
-);
-CREATE INDEX IF NOT EXISTS idx_changesets_created_at ON changesets(created_at);
-CREATE TABLE IF NOT EXISTS changeset_stats (
-    changeset_id   BIGINT NOT NULL REFERENCES changesets(changeset_id),
-    seq_id         BIGINT NOT NULL,
-    uid            BIGINT NOT NULL REFERENCES users(uid),
-    nodes_created  INTEGER DEFAULT 0,
-    nodes_modified INTEGER DEFAULT 0,
-    nodes_deleted  INTEGER DEFAULT 0,
-    ways_created   INTEGER DEFAULT 0,
-    ways_modified  INTEGER DEFAULT 0,
-    ways_deleted   INTEGER DEFAULT 0,
-    rels_created   INTEGER DEFAULT 0,
-    rels_modified  INTEGER DEFAULT 0,
-    rels_deleted   INTEGER DEFAULT 0,
-    poi_created    INTEGER DEFAULT 0,
-    poi_modified   INTEGER DEFAULT 0,
-    tag_stats      JSONB,
-    PRIMARY KEY (seq_id, changeset_id)
-);
-CREATE INDEX IF NOT EXISTS idx_changeset_stats_uid ON changeset_stats(uid);
-CREATE TABLE IF NOT EXISTS state (
-    source_url  TEXT PRIMARY KEY,
-    last_seq    BIGINT NOT NULL,
-    last_ts     TIMESTAMPTZ NOT NULL,
-    updated_at  TIMESTAMPTZ NOT NULL
-);
-"""
 
 _pool: asyncpg.Pool | None = None
